Move the bang algorithm to GroupIterator - the sole user.
